Repo-wide mechanical cleanup: reduce LOC, uniform names, comment hygiene
Mechanical cleanup pass across the project's own code (sycl/, native/, *.js,
tests/, scripts/, .github/workflows/, binding.gyp) — vendored xmrig and
ctpl-stl excluded. Reduce lines of code while keeping readability, make
internal names uniform, delete obvious comments and improve/correct non-obvious
ones, and fix small no-downside issues (dead code/branches, redundant guards,
tab/space, a dead regex capture, stale/incorrect comments) with no behavioral
changes.

Verified on the cleaned tree: Intel B580 7/7 + NVIDIA L4 7/7 GPU hash vectors,
CPU hash vectors 32/32, JS logic 51/51.

+++ sycl/ethash.cpp
@@ -8,7 +8,7 @@
 
 namespace {
 
-constexpr unsigned ETHASH_NODE_WORDS = 16;
+constexpr uint32_t ETHASH_NODE_WORDS = 16;  // 64-byte ethash node = 16 u32 words
 
 union EthashNode {
   uint8_t bytes[ETHASH_NODE_WORDS * sizeof(uint32_t)];
@@ -28,8 +28,9 @@ void keccak512(void* out, const void* in, const unsigned bytes) {
 }
 
 extern "C" ethash_h256_t ethash_get_seedhash(uint64_t epoch) {
+  // Seed hash = keccak256 applied `epoch` times; epoch 0 yields the all-zero hash.
   ethash_h256_t result{};
-  for (uint32_t i = 0; i < epoch; ++i) keccak256(&result, &result, 32);
+  for (uint64_t i = 0; i < epoch; ++i) keccak256(&result, &result, 32);
   return result;
 }
 
@@ -42,11 +43,14 @@ extern "C" bool ethash_compute_cache_nodes(
   const uint32_t num_nodes = static_cast<uint32_t>(cache_size / sizeof(EthashNode));
   auto* nodes = static_cast<EthashNode*>(nodes_ptr);
 
+  // Sequentially seed the cache: each node is keccak512 of the previous one.
   keccak512(nodes[0].bytes, seed, 32);
   for (uint32_t i = 1; i != num_nodes; ++i) {
     keccak512(nodes[i].bytes, nodes[i - 1].bytes, 64);
   }
 
+  // RandMemoHash passes: mix each node with its predecessor (wrapping) XORed
+  // against a pseudo-randomly chosen node, then re-hash in place.
   for (uint32_t j = 0; j != ETHASH_CACHE_ROUNDS; ++j) {
     for (uint32_t i = 0; i != num_nodes; ++i) {
       const uint32_t idx = nodes[i].words[0] % num_nodes;
@@ -58,6 +62,7 @@ extern "C" bool ethash_compute_cache_nodes(
     }
   }
 
+  // Normalize to little-endian word order (no-op on LE hosts) for reproducibility.
   fix_endian_arr32(nodes[0].words, num_nodes * ETHASH_NODE_WORDS);
   return true;
 }
